+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.0.2 (9/9/19)
+
+- Add to all options, visible and predicate
+- Add fieldSet function
+
 ## 1.0.1 (9/9/19)
 
 - Security updates